LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/referrer-policy-from-rules.https_3-3-expected.txt


FAIL with "no-referrer" referrer policy in rule set overriding "unsafe-url" of cross-site referring page assert_equals: must send no referrer expected (undefined) undefined but got (string) "https://localhost:9443/speculation-rules/prefetch/resources/executor.sub.html?uuid=c591f50c-d4e4-4861-b5ec-93480bb49b11"
